+++ test/req/integration_test.exs
@@ -31,7 +31,7 @@ defmodule Req.IntegrationTest do
       encode_body: 1,
       put_params: 1,
       put_range: 1,
-      put_if_modified_since: 1,
+      cache: 1,
       decompress: 1,
       decode_body: 1
     ]